api folder

+++ ecommerce/api/views.py
@@ -1,3 +1,4 @@
+import json
 from django.shortcuts import render
 from rest_framework.views import APIView
 from rest_framework.permissions import IsAuthenticated
@@ -11,6 +12,7 @@ from .serializers import *
 from rest_framework.decorators import permission_classes
 from django.contrib.auth import authenticate, login,logout
 from rest_framework.parsers import MultiPartParser, FormParser
+from django.shortcuts import render, get_object_or_404
 
 
 # Create your views here.
@@ -230,8 +232,43 @@ class CartAPIView(APIView):
 
     def delete(self,request,id):
         print(id)
-        Cart.objects.get(name=id).delete()
-        return Response({'success':'Product deleted successfully'},status=status.HTTP_200_OK)
+        if Cart.objects.filter(name=id).exists():
+            Cart.objects.get(name=id).delete()
+            return Response({'success':'Product deleted successfully'},status=status.HTTP_200_OK)
+        else:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+    
+
+@permission_classes([IsAuthenticated])
+class Track_Order_APIVIEW(APIView):
+    def post(request):
+        if request.method == "POST":
+            order_number = request.POST.get("order_number")
+            order = get_object_or_404(Order, order_number=order_number)
+            return Response({"order": order})
+        return Response( status=status.HTTP_400_BAD_REQUEST)
+    
+    def order_tracker(request):
+        if request.method=="POST":
+            orderId = request.POST.get('orderId', '')
+            print(orderId)
+            try:
+                order=Order.objects.filter(pk=orderId)
+                orderData = get_object_or_404(Order, order_number=order)
+
+                if len(order)>0:
+                    update = Order.objects.filter(pk=orderId)
+                    updates = []
+                    for order in update:
+                        # change order status to scheduled
+                        if order.status == 'processing':
+                            order.status = 'scheduled'
+                            order.save()
+                        updates.append({'status' : order.status})
+                        response = json.dumps(updates)
+                        return Response(response)
+            except Exception as e:
+                return Response( status=status.HTTP_400_BAD_REQUEST)
